backend/requirements.txt
tavily-python
langchain-community
langchain 
langchain_chroma
langchain-groq
langchain-openai
langchainhub
python-dotenv
